requirements.txt
python-dateutil>=2.8.2
python-dotenv>=0.19.0
google-generativeai>=0.8.3
openpyxl>=3.1.5
mysql-connector-python>=9.1.0
pandas>=2.2.3